requirements.txt
streamlit
torch
transformers
langchain
requests
python-dotenv
